[MIG] partner_firstname: Style improvements for forward port * Prefer odoo to openerp & remove data tags * Move exceptions file out of models
