+++ partner_firstname/README.rst
@@ -47,7 +47,7 @@ supply the *last name* and *first name* (just once per contact).
 
 .. image:: https://odoo-community.org/website/image/ir.attachment/5784_f2813bd/datas
    :alt: Try me on Runbot
-   :target: https://runbot.odoo-community.org/runbot/134/8.0
+   :target: https://runbot.odoo-community.org/runbot/134/9.0
 
 For further information, please visit:
 
@@ -68,8 +68,7 @@ Bug Tracker
 Bugs are tracked on `GitHub Issues
 <https://github.com/OCA/partner-contact/issues>`_. In case of trouble, please
 check there if your issue has already been reported. If you spotted it first,
-help us smashing it by providing a detailed and welcomed feedback `here
-<https://github.com/OCA/partner-contact/issues/new?body=module:%20partner_firstname%0Aversion:%208.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
+help us smashing it by providing a detailed and welcomed feedback.
 
 Credits
 =======
@@ -93,6 +92,7 @@ Contributors
 * Ronald Portier <ronald@therp.nl>
 * Sylvain Van Hoof
 * Pedro Baeza <pedro.baeza@serviciosbaeza.com>
+* Dave Lasley <dave@laslabs.com>
 
 Translations
 ------------
