+++ partner_firstname/exceptions.py
