+++ partner_firstname/models/__init__.py
@@ -4,3 +4,4 @@
 
 from . import base_config_settings
 from . import res_partner
+from . import res_users
