+++ partner_firstname/models/res_users.py
