+++ partner_firstname/tests/base.py
@@ -3,7 +3,7 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
 from openerp.tests.common import TransactionCase
-from ..models import exceptions as ex
+from .. import exceptions as ex
 
 
 class MailInstalled():
