+++ partner_firstname/tests/test_empty.py
@@ -8,7 +8,7 @@ To have more accurate results, remove the ``mail`` module before testing.
 """
 from openerp.tests.common import TransactionCase
 from .base import MailInstalled
-from ..models import exceptions as ex
+from .. import exceptions as ex
 
 
 class CompanyCase(TransactionCase):
