+++ partner_firstname/views/res_partner.xml
@@ -1,6 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<openerp>
-<data>
+<odoo>
 
 <record id="view_partner_simple_form_firstname" model="ir.ui.view">
     <field name="name">Add firstname and lastname</field>
@@ -88,5 +87,4 @@
   </field>
 </record>
 
-</data>
-</openerp>
+</odoo>
