+++ partner_firstname/views/res_user.xml
@@ -1,6 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<openerp>
-<data>
+<odoo>
 
 <record id="view_users_form" model="ir.ui.view">
     <field name="name">Add firstname and surnames</field>
@@ -25,5 +24,4 @@
   </field>
 </record>
 
-</data>
-</openerp>
+</odoo>
